rudimentary multi-tube string support
digiline_send(channel, "some string") will now echo "some string"
to all alphanumeric tubes in a lineup, so long as they all face
the same way, starting from the nixie the LuaC is connected to,
going left to right.  The other tubes in the line do not need to
be connected to anything - think of them as being connected
toghether internally.

The string will spread until either a tube is found that faces the
wrong way, or it has a channel that's not empty/nil and is set to
something other than what the first is set to, or if a node is
encountered that is not an alpha-numeric tube at all.

Tubes to the left of the connected one are ignored; you can put
multiple lines of tubes side by side, so long as each tube that
starts each line has a unique channel.

The string is padded with spaces and trimmed to 64 chars.

Special messages:

"off_multi" turns all tubes in the lineup off
"allon_multi" turns on all anode segments of all tubes in the lineup.

Note that the existing messages "get" and "getstr" will only read
the one character from the first, connected tube, not the others
in the lineup.